Added intptr_t casts in comparisons in R_CheckStack for Solaris cc.
git-svn-id: https://svn.r-project.org/R/trunk@63454 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/include/Rinternals.h
@@ -1190,7 +1190,7 @@ void R_Reprotect(SEXP, PROTECT_INDEX);
 	void R_SignalCStackOverflow(void);				\
 	int dummy;							\
 	intptr_t usage = R_CStackDir * (R_CStackStart - (uintptr_t)&dummy); \
-	if(R_CStackLimit != -1 && usage > /*0.95 * */R_CStackLimit)	\
+	if(R_CStackLimit != -1 && usage > ((intptr_t) R_CStackLimit))	\
 	    R_SignalCStackOverflow();					\
     } while (FALSE)
 #endif
